% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_key.R
\name{check_api_key}
\alias{check_api_key}
\title{check purple air api key}
\usage{
check_api_key(purple_air_api_key = Sys.getenv("PURPLE_AIR_API_KEY"))
}
\arguments{
\item{purple_air_api_key}{A character that is your PurpleAir API \code{READ} key}
}
\value{
if the key is valid, a message is emitted and the input is invisibly returned;
invalid keys will throw an R error which utilizes information from the underlying http error
to inform the user
}
\description{
Use the PurpleAir API to validate your Purple Air API Key.
Find more details on this function at https://api.purpleair.com/#api-keys-check-api-key
}
\examples{
check_api_key()
try(check_api_key("foofy"))
}
