% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{find_local_pam}
\alias{find_local_pam}
\title{find purple area monitors on a local network}
\usage{
find_local_pam(
  network_prefix = "192.168.1",
  timeout = 1,
  find_multiple = FALSE
)
}
\arguments{
\item{network_prefix}{character string; base IPv4 prefix
(first three octets) used to generate IP addresses}

\item{timeout}{numeric; number of seconds to wait for each ping}
}
\value{
a list of purple air monitor IDs named according to their IP address
}
\description{
All IP addresses within the network are pinged
to possibly return a purple air monitor sensor ID.
}
\examples{
\dontrun{
find_local_pam()
}
}
