% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_data.R, R/get_sensor_history.R
\name{get_sensor_data}
\alias{get_sensor_data}
\title{get sensor data}
\usage{
get_sensor_data(
  sensor_index,
  fields,
  start_timestamp,
  end_timestamp,
  average = c("real-time", "10min", "30min", "60min", "6hr", "1d", "1w", "1m", "1y"),
  purple_air_api_key = Sys.getenv("PURPLE_AIR_API_KEY"),
  read_key = NULL
)

get_sensor_data(
  sensor_index,
  fields,
  start_timestamp,
  end_timestamp,
  average = c("real-time", "10min", "30min", "60min", "6hr", "1d", "1w", "1m", "1y"),
  purple_air_api_key = Sys.getenv("PURPLE_AIR_API_KEY"),
  read_key = NULL
)
}
\arguments{
\item{sensor_index}{Integer (or numeric, character object coerceable to integer) \code{sensor_index}}

\item{fields}{A character vector of which 'sensor data fields' to return}

\item{start_timestamp}{time stamp of first required history entry (inclusive)}

\item{end_timestamp}{end time stamp of history to return (exclusive)}

\item{average}{time frame to request averaged results for}

\item{purple_air_api_key}{A character that is your PurpleAir API \code{READ} key}

\item{read_key}{A character key required to read data from private devices}
}
\value{
a list of sensor data, named by the provided \code{fields}

a list of sensor data, named by the provided \code{fields}
}
\description{
Retrieves the latest data of a single sensor matching the provided \code{sensor_index}.
Find more details on sensor fields at https://api.purpleair.com/#api-sensors-get-sensor-data.

Retrieves the latest history of a single sensor matching the provided \code{sensor_index}.
Find more details on sensor fields at https://api.purpleair.com/#api-sensors-get-sensor-history.
}
\examples{
get_sensor_data(sensor_index = 175413, fields = c("name", "last_seen", "pm2.5_cf_1", "pm2.5_atm"))
get_sensor_data(sensor_index = "175413", fields = c("name", "last_seen", "pm2.5_cf_1", "pm2.5_atm"))
get_sensor_data(sensor_index = 175413, fields = c("name", "last_seen", "pm2.5_cf_1", "pm2.5_atm"))
get_sensor_data(sensor_index = "175413", fields = c("name", "last_seen", "pm2.5_cf_1", "pm2.5_atm"))
}
