% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{ip_pam_id}
\alias{ip_pam_id}
\title{get purple air monitor id from ip address}
\usage{
ip_pam_id(ip_address, timeout = 1)
}
\arguments{
\item{ip_address}{character; address to send sensor id request to}

\item{timeout}{numeric; number of seconds to wait for each ping}
}
\value{
NULL if address doesn't respond to a SensorId request;
if ip address is a purple air monitor, the monitor id is returned
}
\description{
get purple air monitor id from ip address
}
\examples{
ip_pam_id("192.168.1.144") # purple air
ip_pam_id("192.168.1.148") # no server
ip_pam_id("192.168.1.141") # non-purple air
}
