% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{local_sensor_data}
\alias{local_sensor_data}
\title{local_sensor_data}
\usage{
local_sensor_data(ip_address)
}
\arguments{
\item{ip_address}{address of purple air monitor on local area network
to request data from}
}
\value{
a list of data returned by the sensor
}
\description{
Get latest data (updated every two minutes) from a sensor the local area network.
}
\examples{
\dontrun{
local_sensor_data("192.168.1.144") |>
  _[c("DateTime", "current_temp_f", "current_humidity", "pm2_5_cf_1", "p25aqic")]
}
}
