% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organization_data.R
\name{get_organization_data}
\alias{get_organization_data}
\title{Get Organization Data}
\usage{
get_organization_data(purple_air_api_key = Sys.getenv("PURPLE_AIR_API_KEY"))
}
\arguments{
\item{purple_air_api_key}{A character that is your PurpleAir API \code{READ} key}
}
\value{
A list of organization info
}
\description{
Use the PurpleAir API to retrieve information for the organization containing the provided api_key
Find more details on this function at https://api.purpleair.com/#api-organization-get-organization-data
}
\examples{
\dontrun{
get_organization_data()
}
}
\seealso{
check_api_key
}
