% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_data.R
\name{get_sensor_data}
\alias{get_sensor_data}
\title{get sensor data}
\usage{
get_sensor_data(
  purple_air_api_key = Sys.getenv("PURPLE_AIR_API_KEY"),
  sensor_index,
  fields
)
}
\arguments{
\item{sensor_index}{The \code{sensor_index} as found in the JSON for this specific sensor}

\item{fields}{A character vector of which 'sensor data fields' to return}

\item{purple_air_key}{Your PurpleAir API \code{READ} key}

\item{read_key}{A character key required to read data from private devices}
}
\value{
a list of sensor data, named by the provided \code{fields}
}
\description{
Retrieves the latest data of a single sensor matching the provided \code{sensor_index}.
Find more details on sensor fields at https://api.purpleair.com/#api-sensors-get-sensor-data.
}
\examples{
get_sensor_data(sensor_index = as.integer(175413), fields = c("name", "last_seen", "pm2.5_cf_1", "pm2.5_atm"))
}
