% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensors_data.R
\name{get_sensors_data}
\alias{get_sensors_data}
\title{get sensors data}
\usage{
get_sensors_data(
  x,
  fields,
  location_type = c("both", "inside", "outside"),
  max_age = as.integer(604800),
  purple_air_api_key = Sys.getenv("PURPLE_AIR_API_KEY"),
  read_keys = NULL
)
}
\arguments{
\item{x}{an input object used to define multiple sensors
\itemize{
\item st_bbox object will select sensors geographically (\code{nwlat}, \code{nwlon}, \code{selat}, \code{selon})
\item integer (or numeric or character?) vector will select sensors based on \code{sensor_index} (\code{show_only})
\item POSIXct object will select sensors modified since then (\code{modified_since})
}}

\item{fields}{A character vector of which 'sensor data fields' to return}

\item{location_type}{character; restrict to only "outside" or "inside" sensors (Outside: 0, Inside: 1)}

\item{max_age}{integer; filter results to only include sensors modified or updated within the last number of seconds}

\item{purple_air_api_key}{Your PurpleAir API \code{READ} key}

\item{read_keys}{TODO A character vector of keys required to read data from private devices}
}
\value{
a list of sensor data, named by the provided \code{fields}
}
\description{
Retrieves the latest data of multiple sensors matching the provided parameters.
Find more details on sensor fields at https://api.purpleair.com/#api-sensors-get-sensors-data.
}
\examples{
get_sensors_data(x = as.integer(c(175257, 175413)),
  fields = c("name", "last_seen", "pm2.5_cf_1", "pm2.5_atm"))
get_sensors_data(x = as.integer(c(175257, 175413)),
  fields = c("name", "pm2.5_cf_1", "pm2.5_atm"), location_type = "outside")
}
