% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{local_sensor_live}
\alias{local_sensor_live}
\title{local_sensor_live}
\usage{
local_sensor_live(ip_address)
}
\arguments{
\item{ip_address}{address of purple air monitor on local area network
to request data from}
}
\description{
Stream the latest data from a sensor on the local area network.
Data is updated every second on the device; this function waits
half a second after each call, which takes less than half a second,
ensuring sub one second updating freqency.
}
\examples{
\dontrun{
local_sensor_live("192.168.1.144")
}
}
